requirements.txt
streamlit
tensorflow==2.15
pillow
numpy